third_party/WebKit/LayoutTests/external/wpt/css/selectors/invalidation/matches-expected.txt

This is a testharness.js-based test.
PASS Preconditions.
PASS Invalidate :matches() for simple selector arguments.
PASS Invalidate :matches() for compound selector arguments.
PASS Invalidate :matches() for complex selector arguments.
FAIL Invalidate nested :matches(). assert_equals: expected "rgb(0, 0, 0)" but got "rgb(255, 0, 0)"
PASS Test specificity of :matches().
Harness: the test ran to completion.
